+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 
   - Fixed issue in `find_parameters()` and `get_parameters()` for *gamlss* models.
+  - Fixed issue in `get_data()` for *plm* models, where the `index`-argument was used in the `plm()`-function call.
 
 # insight 0.5.0